% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_functions.R
\name{pgdpois}
\alias{pgdpois}
\title{Cumulative Distribution Function of the GD-Poisson Distribution}
\usage{
pgdpois(q, lambda, theta)
}
\arguments{
\item{q}{A vector of non-negative integers at which to evaluate the CDF.}

\item{lambda}{A positive parameter representing the mean of the distribution.}

\item{theta}{A positive parameter representing the dispersion of the distribution.}
}
\value{
A numeric vector of the same length as \code{q}, containing the CDF evaluated at each element of \code{q}.
}
\description{
Computes the cumulative distribution function (CDF) of the GD-Poisson distribution for a vector of integer values. The parameters `lambda` and `theta` can also be vectors of the same length as `q`, or any parameter can be a single value to be applied across all evaluations.
}
\details{
The GD-Poisson distribution is a generalization of the Poisson distribution that introduces arbitrary underdispersion.
}
\examples{
# Evaluate the CDF at q = 0:10 with lambda = 5 and theta = 4
pgdpois(0:10, lambda = 5, theta = 6)

}
