% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_functions.R
\name{qgdpois}
\alias{qgdpois}
\title{Quantile Function of the GD-Poisson Distribution}
\usage{
qgdpois(p, lambda, theta, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{A vector of probabilities (each between 0 and 1) at which to evaluate the quantiles.}

\item{lambda}{A positive parameter representing the mean of the distribution.}

\item{theta}{A positive parameter representing the dispersion of the distribution.}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities are \eqn{P(X \leq x)}, otherwise, \eqn{P(X > x)}.}

\item{log.p}{Logical; if \code{TRUE}, probabilities \code{p} are given as \code{log(p)}.}
}
\value{
An integer vector of the same length as \code{p}, containing the quantiles corresponding to each probability in \code{p}.
}
\description{
Computes the quantile function of the GD-Poisson distribution for a vector of probabilities. The parameters \code{lambda} and \code{theta} can also be vectors of the same length as \code{p}, or any parameter can be a single value to be applied across all evaluations.
}
\details{
The quantile function finds the smallest integer \eqn{k} such that the CDF at \eqn{k} is greater than or equal to the specified probability \eqn{p}.

The function is vectorized, meaning that \code{p}, \code{lambda}, and \code{theta} can be vectors. If any of these parameters are single values (length 1), they will be used for all evaluations. All vector inputs must either be of the same length or have a length of 1.
}
\examples{
# Find the 25th, 50th, and 75th percentiles with lambda = 2 and theta = 1
qgdpois(c(0.25, 0.5, 0.75), lambda = 10, theta = 1)

# Vectorized parameters: different values for theta
qgdpois(c(0.25, 0.5, 0.75), lambda = c(1, 5.3, 6.5), theta = 8)

# Scalar values
qgdpois(0.99, lambda = 24, theta = 1)

}
