% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_functions.R
\name{rgdpois}
\alias{rgdpois}
\title{Random Generation from the GD-Poisson Distribution}
\usage{
rgdpois(n, lambda, theta)
}
\arguments{
\item{n}{Integer specifying the number of random samples to generate.}

\item{lambda}{A positive parameter representing the mean of the distribution.}

\item{theta}{A positive parameter representing the dispersion of the distribution.}
}
\value{
An integer vector of length \code{n}, containing random samples from the GD-Poisson distribution.
}
\description{
Generates random samples from the GD-Poisson distribution.
}
\details{
The generated random sample follow the GD-Poisson distribution.
}
\examples{
# Generate 10 random samples with lambda = 2 and theta = 1
rgdpois(10, lambda = 2, theta = 1)
}
